+++ app/src/main/java/application/repository/JogoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import application.model.Jogo;
 
-public interface JogoRepository extends CrudRepository<Jogo, Long> {
+public interface JogoRepository extends CrudRepository<Jogo, Integer> {
     public Iterable<Jogo> findByTitulo(String titulo);
 }
